+++ example/lib/injector.iconfig.dart
@@ -4,33 +4,17 @@
 // InjectableConfigGenerator
 // **************************************************************************
 
-import 'package:example/services.dart';
-import 'package:dio/dio.dart';
-import 'package:example/register_module.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:example/register_module.dart';
 import 'package:get_it/get_it.dart';
 
 Future<void> $initGetIt(GetIt g, {String environment}) async {
-  final registerModule = _$RegisterModule(g);
-  g.registerFactory<Service11>(() => Service11());
-  g.registerFactory<ServiceAbs>(() => registerModule.serviceAA);
-
-  //Register dev Dependencies --------
-  if (environment == 'dev') {
-    final sharedPreferences = await registerModule.prefs;
-    g.registerFactory<SharedPreferences>(() => sharedPreferences);
-  }
-
-  //Eager singletons must be registered in the right order
-  g.registerSingleton<Dio>(registerModule.dioDev);
+  final registerModule = _$RegisterModule();
+  final sharedPreferences = await registerModule.prefs;
+  g.registerFactory<SharedPreferences>(() => sharedPreferences);
+  g.registerFactoryParam<BackendService, String, dynamic>(
+      (url, _) => registerModule.getService(url));
+  g.registerFactoryAsync<ApiClient>(() => ApiClient.create());
 }
 
-class _$RegisterModule extends RegisterModule {
-  final GetIt _g;
-  _$RegisterModule(this._g);
-  ServiceAA get serviceAA => ServiceAA(
-        _g<FirebaseAuth>(),
-        _g<Dio>(),
-      );
-}
+class _$RegisterModule extends RegisterModule {}