+++ example/lib/register_module.dart
@@ -1,23 +1,41 @@
-import 'package:dio/dio.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:injectable/injectable.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 @registerModule
 abstract class RegisterModule {
-  // @dev
-  @singleton
-  Dio get dioDev => Dio(BaseOptions(baseUrl: "basweUwrl"));
-
-  @RegisterAs(ServiceAbs)
-  ServiceAA get serviceAA;
-
-  @dev
+  @preResolve
   Future<SharedPreferences> get prefs => SharedPreferences.getInstance();
+  // AbsService get absS;
+  // Future<ServiceX> genIml(String x, int y) async => ServiceX();
+  // @singleton
+  // ServiceX ser();
+  // @singleton
+  // TestClass get testClass;
+  // TestSingleton2 get testSing33;
+
+  // TestSingleton testSing(String x, int y) => TestSingleton();
+
+  BackendService getService(String url) => BackendService(url);
+  // @preResolve
+  // @Singleton(dependsOn: [TestClass])
+  // Future<TestSingleton> get futureSing => TestSingleton.create();
 }
 
-class ServiceAA implements ServiceAbs {
-  ServiceAA(FirebaseAuth auth, Dio dio);
+// @injectable
+// class TestSingleton2 {
+//   TestSingleton2(@factoryParam String x, TestSingleton test);
+// }
+
+abstract class AbsService<T> {}
+
+class BackendService extends AbsService<int> {
+  BackendService(@factoryParam String url);
 }
 
-abstract class ServiceAbs {}
+@injectable
+class ApiClient {
+  @factoryMethod
+  static Future<ApiClient> create() async {
+    return ApiClient();
+  }
+}