+++ injectable_generator/lib/injectable_generator.dart
@@ -36,15 +36,18 @@ class InjectableGenerator implements Generator {
       if (moduleChecker.hasAnnotationOfExact(clazz)) {
         throwBoxedIf(
             !clazz.isAbstract, '[${clazz.name}] must be an abstract class!');
-
-        for (var annotatedElement in clazz.accessors) {
-          allDepsInStep.add(await DependencyResolver(annotatedElement)
-              .resolveFromAccessor(clazz, buildStep.resolver));
+        final executables = <ExecutableElement>[
+          ...clazz.accessors,
+          ...clazz.methods,
+        ];
+        for (var annotatedElement in executables) {
+          allDepsInStep.add(await DependencyResolver(buildStep.resolver)
+              .resolveModuleMemeber(clazz, annotatedElement));
         }
       } else if (_hasInjectable(clazz) ||
           (autoRegister && _hasConventionalMatch(clazz))) {
         allDepsInStep
-            .add(await DependencyResolver(clazz).resolve(buildStep.resolver));
+            .add(await DependencyResolver(buildStep.resolver).resolve(clazz));
       }
     }