+++ injectable_generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # ChangeLog
+
+## [0.3.0] Breaking Changes!
+- add support for GetIt 4.0.0
+- fix generic types are registered as dynamic
+- fix unresolved future when registering asynchronous dependencies
+- change asynchronous dependencies will be registered using async factory unless annotated with @preResolve
+
+## [0.2.4]
+
+- fix type arguments are not imported.
+- fix get with instanceNname missing type.
+
 ## [0.2.3]
+
 - improve support for custom initializers
 - fix src import issue
 
@@ -8,8 +21,6 @@
 - remove flutter dependency
 - add support for custom initializers in register Modules
 - minor fixes
-
-# ChangeLog
 
 ## [0.2.1]